Refactor application structure and add new files
Updated project files to include new source and header files: `App.cpp`, `App.h`, `MainFrame.cpp`, and `MainFrame.h`. Introduced filters for better organization. Refactored `main.cpp` to remove class definitions and include `App.h`, with the `OnInit` method now implemented in `App.cpp`. This restructuring enhances maintainability and clarity of the application logic.
